Fix: Fixed duplication bug with usePresence
