+++ client/src/customHooks/usePresence.js
@@ -2,27 +2,82 @@ import { useEffect, useState } from "react";
 import { socket } from "../socket";
 
 const usePresence = (channelName, initialUserInfo) => {
-  const [userInfo] = useState(initialUserInfo);
   const [currentPresence] = useState(channelName);
+  const [presenceData, setPresenceData] = useState([]);
+  const [userInfo] = useState(initialUserInfo);
 
   useEffect(() => {
     if (!currentPresence) {
       return;
     }
 
-    socket.emit("presenceSet:enter", currentPresence, userInfo);
+    const handlePresenceLeave = (socketId) => {
+      setPresenceData((prevData) =>
+        prevData.filter((data) => data.socketId !== socketId)
+      );
+    };
+
+    const handlePresenceJoin = (response) => {
+      setPresenceData((prevData) => [...prevData, response]);
+    };
+
+    const handlePresenceUpdate = (response) => {
+      setPresenceData((prevData) =>
+        prevData.map((data) =>
+          data.socketId === response.socketId ? response : data
+        )
+      );
+    };
+
+    const handlePresenceSubscribe = (ack) => {
+      if (ack.success === true) {
+        setPresenceData(ack.users);
+        socket.off(`presence:${currentPresence}:leave`, handlePresenceLeave);
+        socket.off(`presence:${currentPresence}:join`, handlePresenceJoin);
+        socket.off(`presence:${currentPresence}:update`, handlePresenceUpdate);
+
+        socket.on(`presence:${currentPresence}:leave`, handlePresenceLeave);
+        socket.on(`presence:${currentPresence}:join`, handlePresenceJoin);
+        socket.on(`presence:${currentPresence}:update`, handlePresenceUpdate);
+
+        // Enter user into presence set
+        socket.emit("presenceSet:enter", currentPresence, userInfo);
+      }
+    };
+
+    const handleSocketConnect = () => {
+      // Re-subscribe on socket reconnect
+      console.log(currentPresence, "current presence");
+      socket.emit(
+        "presence:subscribe",
+        currentPresence,
+        handlePresenceSubscribe
+      );
+    };
+
+    socket.emit("presence:subscribe", currentPresence, handlePresenceSubscribe);
+    socket.on("connect", handleSocketConnect);
 
     return () => {
       socket.emit("presenceSet:leave", currentPresence);
-      console.log("user has left presence set");
+      socket.emit(`presence:unsubscribe`);
+      socket.off("connect", handleSocketConnect);
+
+      socket.off(`presence:${currentPresence}:leave`, handlePresenceLeave);
+      socket.off(`presence:${currentPresence}:join`, handlePresenceJoin);
+      socket.off(`presence:${currentPresence}:update`, handlePresenceUpdate);
     };
-  }, [currentPresence, userInfo]); // Add dependencies
+  }, [currentPresence, userInfo]);
+
+  useEffect(() => {
+    console.log("Presence Data Updated:", presenceData);
+  }, [presenceData]);
 
   const updateUserInfo = async (updatedUserInfo) => {
     socket.emit("presence:update", currentPresence, updatedUserInfo);
   };
 
-  return { updateUserInfo };
+  return { presenceData, updateUserInfo };
 };
 
 export default usePresence;
